fix(sw360adapter): refactoring to prvent UnsupportedOperationException
