+++ antenna-sw360-adapter/src/main/java/org/eclipse/sw360/antenna/sw360/adapter/commonComparisonProperties/ArtifactCommons.java
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.sw360.antenna.sw360.comparator;
+package org.eclipse.sw360.antenna.sw360.adapter.commonComparisonProperties;
 
 import org.eclipse.sw360.antenna.model.artifact.Artifact;
 import org.eclipse.sw360.antenna.sw360.utils.SW360ComponentAdapterUtils;
