--- antenna-sw360-adapter/src/main/java/org/eclipse/sw360/antenna/sw360/comparator/ICommonComparisonProperties.java
@@ -1,19 +1,0 @@
-/*
- * Copyright (c) Bosch Software Innovations GmbH 2018.
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v2.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v20.html
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.eclipse.sw360.antenna.sw360.comparator;
-
-import java.util.List;
-
-public interface ICommonComparisonProperties {
-    String getName();
-
-    List<String> getVersions();
-}
